backend/requirements.txt
Flask
Flask-Cors
Flask-PyMongo
PyJWT
python-dotenv
google-generativeai